문자 중계 개선
