+++ robocup-web/src/components/matches/match/Match.vue
@@ -77,71 +77,7 @@
       </v-flex>
     </v-layout>
 
-    <v-form ref="eventForm" v-model="valid" v-show="isSignIn">
-      <v-layout row wrap mt-5>
-        <v-flex offset-md2 sm2 md2 lg3 xl3>
-          <v-select
-            ref="selectEventType"
-            v-model="eventType"
-            :items="eventTypes"
-            label="Type"
-            color="cyan"
-            outline
-          ></v-select>
-        </v-flex>
-
-        <v-flex ml-3>
-          <v-text-field
-            ref="tfEvent"
-            v-model="eventText"
-            label="Event"
-            required
-            :rules="baseRules"
-          />
-        </v-flex>
-
-        <v-flex ml-3>
-          <v-btn fab dark color="cyan" small outline @click="saveEvent">
-            <v-icon dark>add</v-icon>
-          </v-btn>
-        </v-flex>
-      </v-layout>
-    </v-form>
-
-    <v-layout row mt-5>
-      <v-flex offset-md2>
-        <v-timeline align-top dense v-for="event in events" :key="event.time" xs3 pa-1>
-          <v-timeline-item color="teal lighten-3" small>
-            <v-layout wrap pt-3>
-              <v-flex xs3>
-                <span style="color: gray">{{event.time}}"</span>
-              </v-flex>
-              <v-flex ml-2>
-                <strong>{{event.text}}</strong>
-                <div class="caption mb-2">{{event.type}}</div>
-                <v-avatar v-show="event.type == 'GOAL'">
-                  <v-img :src="require('@/assets/icon_goal.png')"></v-img>
-                </v-avatar>
-                <v-avatar v-show="event.type == 'FOUL'">
-                  <v-img :src="require('@/assets/icon_red_card.png')"></v-img>
-                </v-avatar>
-              </v-flex>
-              <v-btn
-                v-show="isSignIn"
-                fab
-                dark
-                small
-                outline
-                color="cyan"
-                @click="deleteEvent(event)"
-              >
-                <v-icon dark>remove</v-icon>
-              </v-btn>
-            </v-layout>
-          </v-timeline-item>
-        </v-timeline>
-      </v-flex>
-    </v-layout>
+    <ReportContainer :matchId="matchId" :teamAId="teamAId" :teamBId="teamBId"></ReportContainer>
     <CommentContainer :type="collectionName" :parentId="matchId"></CommentContainer>
   </v-container>
 </template>
@@ -150,15 +86,13 @@
 <script>
 import CommentContainer from "../../comments/CommentContainer";
 import VideoLinkDialog from "./VideoLinkDialog";
-
-const EVENT_TYPE_GENERAL = "GENERAL";
-const EVENT_TYPE_GOAL = "GOAL";
-const EVENT_TYPE_FOUL = "FOUL";
+import ReportContainer from "./report/ReportContainer";
 
 export default {
   components: {
     CommentContainer,
-    VideoLinkDialog
+    VideoLinkDialog,
+    ReportContainer,
   },
   data: () => ({
     collectionName: "matches",
@@ -177,21 +111,14 @@ export default {
     type: "Friendly Match",
     matchDate: "",
     matchTime: "",
+    teamAId: "",
     teamAName: "",
     teamAScore: 0,
     teamALogo: "icon_help.png",
+    teamBId: "",
     teamBName: "",
     teamBScore: 0,
     teamBLogo: "icon_help.png",
-    baseRules: [
-      v => !!v || "required",
-      v => (v && v.length <= 50) || "Text must be less than 50 characters"
-    ],
-    events: [],
-    eventTypes: [EVENT_TYPE_GENERAL, EVENT_TYPE_GOAL, EVENT_TYPE_FOUL],
-    eventType: EVENT_TYPE_GENERAL,
-    eventText: "",
-    valid: true,
     facebookVideoUrl: ""
   }),
   created() {
@@ -213,36 +140,15 @@ export default {
         this.matchTime = `${d.getHours()}:${("0" + d.getMinutes()).slice(-2)}`;
         this.stadium = match.stadium;
         this.type = match.type;
+        this.teamAId = match.teamAId;
         this.teamAName = match.teamAName;
         this.teamAScore = match.teamAScore;
         this.teamALogo = match.teamALogo;
+        this.teamBId = match.teamBId;
         this.teamBName = match.teamBName;
         this.teamBScore = match.teamBScore;
         this.teamBLogo = match.teamBLogo;
         this.facebookVideoUrl = match.facebookVideoUrl;
-      });
-
-    let events = this.events;
-    firebase
-      .firestore()
-      .collection("matches")
-      .doc(this.matchId)
-      .collection("events")
-      .onSnapshot(snapshot => {
-        snapshot.docChanges().forEach(change => {
-          if (change.type === "removed") {
-            const event = change.doc.data();
-            const idx = events.findIndex(function(e) {
-              return e.time === event.time;
-            });
-            if (idx > -1) events.splice(idx, 1);
-          } else {
-            const event = change.doc.data();
-            events.push(event);
-            this.eventType = EVENT_TYPE_GENERAL;
-            this.eventText = "";
-          }
-        });
       });
   },
   updated() {
@@ -344,46 +250,6 @@ export default {
       this.teamBScore--;
       const score = { teamBScore: this.teamBScore };
       updateScore(score, this.matchId);
-    },
-    saveEvent() {
-      if (this.$refs.eventForm.validate()) {
-        const d = new Date();
-        const eventTime = `${d.getHours()}:${("0" + d.getMinutes()).slice(
-          -2
-        )}:${("0" + d.getSeconds()).slice(-2)}`;
-        const event = {
-          time: eventTime,
-          text: this.eventText,
-          type: this.eventType
-        };
-
-        firebase
-          .firestore()
-          .collection("matches")
-          .doc(this.matchId)
-          .collection("events")
-          .doc(eventTime)
-          .set(event)
-          .then(function() {})
-          .catch(function(error) {
-            console.error("Error writing document: ", error);
-          });
-      }
-    },
-    deleteEvent(event) {
-      firebase
-        .firestore()
-        .collection("matches")
-        .doc(this.matchId)
-        .collection("events")
-        .doc(event.time)
-        .delete()
-        .then(function() {
-          console.log("Document successfully deleted!");
-        })
-        .catch(function(error) {
-          console.error("Error removing document: ", error);
-        });
     },
     addVideo(url) {
       this.facebookVideoUrl = url;
